feat(filter): refactor codebase
