+++ packages/paste-website/src/component-examples/filter-group/helpers.ts
@@ -16,7 +16,6 @@ export const filterBySearchString = (uniqueName: string, sid: string, searchValu
 };
 
 export const filterByRoomType = (roomType: RoomTypes, filterValue: RoomTypes): boolean => {
-  if (filterValue === "All") return true;
   return roomType === filterValue;
 };
